Moved upgrades to a separate package
